feat: completely refactored phase1 version of app where everything is now modular. Confirmed things still work through testing
